requirements.txt
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
flake8==6.1.0
idna==3.4
mccabe==0.7.0
mecab-python3==1.0.8
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
PyAudio==0.2.14
pycodestyle==2.11.1
pyflakes==3.1.0
requests==2.31.0
SpeechRecognition==3.10.0
unidic-lite==1.0.8
urllib3==2.0.7
